chore: refactor changes in #628 to work for LogsVolumePanel
